feat(firebaseServices): Restructure to single-page app rather than one page persection - add 1KeySigNote, 2KeySigText, and 3WriteScales. Move original exam to its own folder and endpoint. Move the home base logic to the exam folder page, with its function named ExamHomePage - set up basics with VIEW_STATES and importing user from useAuthContext.
