+++ app/components/3WriteScales.tsx
