+++ app/context/examContext.tsx
@@ -7,6 +7,7 @@ import { ExamContextType } from "../lib/typesAndInterfaces";
 import ExamContext from "./createExamContext";
 import { useRouter } from "next/navigation";
 import { set } from "firebase/database";
+import { getUserSnapshot } from "@/firebase/firestore/model";
 
 export default function ExamContextProvider({ children }: ExamContextType) {
   const [loading, setLoading] = useState(true);
@@ -17,67 +18,22 @@ export default function ExamContextProvider({ children }: ExamContextType) {
 
   const user = auth.currentUser?.displayName;
 
-  async function getUserSnapshot(currentUser: string) {
-    try {
-      const q = query(collection(db, currentUser));
-      const querySnapshot = await getDocs(q);
-      // console.log("querySnapshot in getUserSnapshot:", querySnapshot);
-
-      const res = querySnapshot.docs.map((doc) => ({
-        id: doc.id,
-        ...doc.data(),
-      }));
-      setFormInput(res[0]);
-      console.log("formInput in getUserSnapshot after res:", formInput);
-      return {
-        success: true,
-        message: `Successfully fetched ${currentUser} data`,
-        res,
-      };
-    } catch (err) {
-      return {
-        success: false,
-        message: `Error fetching ${currentUser} data: ${err}`,
-      };
-    }
-  }
-
   useEffect(() => {
-    getUserSnapshot(user);
+    const fetchSnapshot = async () => {
+      try {
+        const { success, message, error, res } = await getUserSnapshot();
+        if (error) {
+          console.error(message);
+        } else {
+          console.log("res in fetchSnapshot:", res);
+        }
+      } catch (error) {
+        console.error(error);
+      }
+    };
+    fetchSnapshot();
     setLoading(false);
   }, []);
-
-  // async function getUserSnapshot(currentUser: string | null | undefined) {
-  //   try {
-  //     const docRef = doc(db, `${currentUser}`, "formInput");
-  //     const docSnap = await getDoc(docRef);
-  //     if (docSnap.exists()) {
-  //       console.log("Document data:", docSnap.data());
-  //       setFormInput(docSnap.data().formInput);
-  //       // console.log(
-  //       //   "formInput after getUserSnapshot docSnap data:",
-  //       //   formInput
-  //       // );
-  //     } else {
-  //       console.log("No such document! docSnap.data() is undefined");
-  //     }
-  //   } catch (err) {
-  //     console.error("getUserSnapshot error:", err);
-  //   }
-  // }
-
-  // function getAndSetUser() {
-  //   console.log("user in getAndSetUser:", user);
-  //   if (!user) {
-  //     setUser(userRef?.displayName);
-  //   }
-  //   getUserSnapshot(user);
-  // }
-
-  // useEffect(() => {
-  //   getUserSnapshot(userRef?.displayName);
-  //   setLoading(false);
-  // }, []);
 
   return (
     <ExamContext.Provider value={formInput}>
